Added Initial discard.
Initial discard of the whole device should be done before the test
begins. If not in RANDOM_IO mode, only the necessary part of the device
was trimmed. But this was not enough, since we do not know the state of
rest of the device. In the worst case it can be full and not trimmed,
thus degrade performance significantly.