Add debugging macros
Add two debugging macros. If DEBUG_NO_DISCARD is defined then
test-discard will not call BLKDISCARD ioctl() at all. It is useful when
debugging, because we do not actually need to do discard at all.

If DEBUG_NO_PREPARE is defined then test-discard will skip device
preparation prior to testing the device. This is useful when debugging
because preparation takes a lot of time, but we do not need it.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>